Recreating the KeyDaemon to use DaemonFramework.
KeyDaemon was already mostly implemented, until it became clear that a generic
privileged helper daemon framework would be more useful to me, and I stripped it
down to create DaemonFramework. I am recreating the repository because adding a
fork as a submodule is probably impossible and definitely unwise. Previous
KeyDaemon repository history is still preserved in the DaemonFramework
submodule.